Add the bind service and account manager strategies for broker calls. Add the initialization of the strategies. Fix the config merging.
